Moved TaskCompleter to Services folder
